+++ src/blocks/gallery-stacked/edit.js
@@ -13,7 +13,7 @@ import Controls from './controls';
 import GalleryImage from '../../components/block-gallery/gallery-image';
 import GalleryPlaceholder from '../../components/block-gallery/gallery-placeholder';
 import GalleryDropZone from '../../components/block-gallery/gallery-dropzone';
-import { BackgroundStyles } from '../../components/block-gallery/background';
+import { BackgroundStyles } from '../../components/background';
 import { GalleryClasses } from '../../components/block-gallery/shared';
 
 /**
